Refactor Filter and Product sections.
